# [WIP] Adding ISAL optimization for parquet gzip decompression